+++ Source/equipment/sunflare-base.js
@@ -1,7 +1,7 @@
 const Equipment = require('../../Classes/Equipment.js');
 const { addModifier } = require('../combatantDAO.js');
 
-module.exports = new Equipment("Sun Flare", 1, "*Inflict @{mod1Stacks} @{mod1} on a foe (+@{speedBonus} speed bonus)*\nCritical Hit: Inflict @{mod2Stacks} @{mod2} as well", "Fire", effect, ["Evasive Sun Flare", "Swift Sun Flare", "Tormenting Sun Flare"])
+module.exports = new Equipment("Sun Flare", 1, "*Inflict @{mod1Stacks} @{mod1} on a foe (+@{speedBonus} speed bonus)*\nCritical Hit: Inflict @{mod2Stacks} @{mod2} as well", "Light", effect, ["Evasive Sun Flare", "Swift Sun Flare", "Tormenting Sun Flare"])
 	.setCategory("Spell")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Stagger", stacks: 1 }, { name: "Slow", stacks: 2 }])